LongTask Timing_ long task in rAF.html

<!DOCTYPE html>
<!-- saved from url=(0056)http://w3c-test.org/longtask-timing/longtask-in-raf.html -->
<html><head><meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
<title>LongTask Timing: long task in rAF</title>
<style>html {    font-family:DejaVu Sans, Bitstream Vera Sans, Arial, Sans;}#log .warning,#log .warning a {  color: black;  background: yellow;}#log .error,#log .error a {  color: white;  background: red;}section#summary {    margin-bottom:1em;}table#results {    border-collapse:collapse;    table-layout:fixed;    width:100%;}table#results > thead > tr > th:first-child,table#results > tbody > tr > td:first-child {    width:8em;}table#results > thead > tr > th:last-child,table#results > thead > tr > td:last-child {    width:50%;}table#results.assertions > thead > tr > th:last-child,table#results.assertions > tbody > tr > td:last-child {    width:35%;}table#results > thead > > tr > th {    padding:0;    padding-bottom:0.5em;    border-bottom:medium solid black;}table#results > tbody > tr> td {    padding:1em;    padding-bottom:0.5em;    border-bottom:thin solid black;}.pass {    color:green;}.fail {    color:red;}tr.timeout {    color:red;}tr.notrun {    color:blue;}tr.optionalunsupported {    color:blue;}.ok {    color:green;}.error {    color:red;}.pass, .fail, .timeout, .notrun, .optionalunsupported .ok, .timeout, .error {    font-variant:small-caps;}table#results span {    display:block;}table#results span.expected {    font-family:DejaVu Sans Mono, Bitstream Vera Sans Mono, Monospace;    white-space:pre;}table#results span.actual {    font-family:DejaVu Sans Mono, Bitstream Vera Sans Mono, Monospace;    white-space:pre;}</style></head><body>

<script src="./LongTask Timing_ long task in rAF_files/testharness.js.download"></script>
<script src="./LongTask Timing_ long task in rAF_files/testharnessreport.js.download"></script>
<script src="./LongTask Timing_ long task in rAF_files/utils.js.download"></script>

<h1>Long Task: requestAnimationFrame</h1>
<div id="log"><section id="summary"><h2>Summary</h2><section><p>Harness status: <span class="ok">OK</span></p></section><p>Found 1 tests</p><div><div><label><input type="checkbox" checked="checked">1 <span class="pass">Pass</span></label></div></div></section><section><h2>Details</h2><table id="results"><thead><tr><th>Result</th><th>Test Name</th><th>Message</th></tr></thead><tbody><tr class="overall-pass"><td class="pass">Pass</td><td>Performance longtask entries are observable.</td><td> <details><summary>Asserts run</summary><table><tbody><tr class="overall-pass"><td class="pass">Pass</td><td><pre><strong>assert_implements</strong>(function "function PerformanceLongTaskTiming() { [native code] }", "Longtasks are not supported.")
    at Test.&lt;anonymous&gt; ( /longtask-timing/longtask-in-raf.html:14:5)</pre></td></tr>
<tr class="overall-pass"><td class="pass">Pass</td><td><pre><strong>assert_equals</strong>(1, 1, "Exactly one entry is expected.")
    at Test.&lt;anonymous&gt; ( /longtask-timing/longtask-in-raf.html:18:13)</pre></td></tr>
<tr class="overall-pass"><td class="pass">Pass</td><td><pre><strong>assert_equals</strong>("longtask", "longtask")
    at checkLongTaskEntry ( /longtask-timing/resources/utils.js:2:3)</pre></td></tr>
<tr class="overall-pass"><td class="pass">Pass</td><td><pre><strong>assert_equals</strong>("self", "self")
    at checkLongTaskEntry ( /longtask-timing/resources/utils.js:3:3)</pre></td></tr>
<tr class="overall-pass"><td class="pass">Pass</td><td><pre><strong>assert_true</strong>(true)
    at checkLongTaskEntry ( /longtask-timing/resources/utils.js:4:3)</pre></td></tr>
<tr class="overall-pass"><td class="pass">Pass</td><td><pre><strong>assert_greater_than_equal</strong>(60, 50)
    at checkLongTaskEntry ( /longtask-timing/resources/utils.js:5:3)</pre></td></tr>
<tr class="overall-pass"><td class="pass">Pass</td><td><pre><strong>assert_greater_than_equal</strong>(259.9000000022352, 0)
    at checkLongTaskEntry ( /longtask-timing/resources/utils.js:6:3)</pre></td></tr>
<tr class="overall-pass"><td class="pass">Pass</td><td><pre><strong>assert_less_than_equal</strong>(259.9000000022352, 321.4000000022352)
    at checkLongTaskEntry ( /longtask-timing/resources/utils.js:8:3)</pre></td></tr>
<tr class="overall-pass"><td class="pass">Pass</td><td><pre><strong>assert_equals</strong>(1, 1, "Exactly one attribution entry is expected")
    at Test.&lt;anonymous&gt; ( /longtask-timing/longtask-in-raf.html:23:13)</pre></td></tr>
<tr class="overall-pass"><td class="pass">Pass</td><td><pre><strong>assert_equals</strong>("taskattribution", "taskattribution")
    at Test.&lt;anonymous&gt; ( /longtask-timing/longtask-in-raf.html:26:13)</pre></td></tr>
<tr class="overall-pass"><td class="pass">Pass</td><td><pre><strong>assert_equals</strong>("unknown", "unknown")
    at Test.&lt;anonymous&gt; ( /longtask-timing/longtask-in-raf.html:27:13)</pre></td></tr>
<tr class="overall-pass"><td class="pass">Pass</td><td><pre><strong>assert_equals</strong>(0, 0)
    at Test.&lt;anonymous&gt; ( /longtask-timing/longtask-in-raf.html:28:13)</pre></td></tr>
<tr class="overall-pass"><td class="pass">Pass</td><td><pre><strong>assert_equals</strong>(0, 0)
    at Test.&lt;anonymous&gt; ( /longtask-timing/longtask-in-raf.html:29:13)</pre></td></tr>
<tr class="overall-pass"><td class="pass">Pass</td><td><pre><strong>assert_equals</strong>("window", "window")
    at Test.&lt;anonymous&gt; ( /longtask-timing/longtask-in-raf.html:30:13)</pre></td></tr>
<tr class="overall-pass"><td class="pass">Pass</td><td><pre><strong>assert_equals</strong>("", "")
    at Test.&lt;anonymous&gt; ( /longtask-timing/longtask-in-raf.html:31:13)</pre></td></tr>
<tr class="overall-pass"><td class="pass">Pass</td><td><pre><strong>assert_equals</strong>("", "")
    at Test.&lt;anonymous&gt; ( /longtask-timing/longtask-in-raf.html:32:13)</pre></td></tr>
<tr class="overall-pass"><td class="pass">Pass</td><td><pre><strong>assert_equals</strong>("", "")
    at Test.&lt;anonymous&gt; ( /longtask-timing/longtask-in-raf.html:33:13)</pre></td></tr></tbody></table></details></td></tr></tbody></table></section></div>
<script>
  async_test(function (t) {
    assert_implements(window.PerformanceLongTaskTiming, 'Longtasks are not supported.');
    const observer = new PerformanceObserver(
        t.step_func(function (entryList) {
            const entries = entryList.getEntries();
            assert_equals(entries.length, 1,
                'Exactly one entry is expected.');
            const longtask = entries[0];
            checkLongTaskEntry(longtask);
            // Assert the TaskAttributionTiming entry in attribution.
            assert_equals(longtask.attribution.length, 1,
                'Exactly one attribution entry is expected');
            const attribution = longtask.attribution[0];
            assert_equals(attribution.entryType, 'taskattribution');
            assert_equals(attribution.name, 'unknown');
            assert_equals(attribution.duration, 0);
            assert_equals(attribution.startTime, 0);
            assert_equals(attribution.containerType, 'window');
            assert_equals(attribution.containerId, '');
            assert_equals(attribution.containerName, '');
            assert_equals(attribution.containerSrc, '');
            observer.disconnect();
            t.done();
        })
    );
    observer.observe({entryTypes: ['longtask']});
    window.onload = () => {
        const script = document.createElement('script');
        script.src = 'resources/raflongtask.js';
        document.body.appendChild(script);
    };
}, 'Performance longtask entries are observable.');
</script>

<script src="./LongTask Timing_ long task in rAF_files/raflongtask.js.download"></script><script type="text/json" id="__testharness__results__">{"test":"http://w3c-test.org/longtask-timing/longtask-in-raf.html","tests":[{"name":"Performance longtask entries are observable.","status":0,"message":null,"stack":null}],"status":0,"message":null,"stack":null}</script></body></html>
